Make AnimationLayer event-aware
