+++ src/main/java/com/jcloisterzone/ui/grid/MainPanel.java
@@ -6,20 +6,14 @@ import java.awt.Component;
 import javax.swing.JPanel;
 
 import com.google.common.eventbus.Subscribe;
-import com.jcloisterzone.Player;
-import com.jcloisterzone.board.Position;
-import com.jcloisterzone.board.Tile;
 import com.jcloisterzone.event.BridgeEvent;
 import com.jcloisterzone.event.CastleDeployedEvent;
 import com.jcloisterzone.event.CornCirclesOptionEvent;
-import com.jcloisterzone.event.FlierRollEvent;
 import com.jcloisterzone.event.GoldChangeEvent;
 import com.jcloisterzone.event.MeepleEvent;
 import com.jcloisterzone.event.ScoreEvent;
 import com.jcloisterzone.event.TileEvent;
 import com.jcloisterzone.event.TowerIncreasedEvent;
-import com.jcloisterzone.feature.Feature;
-import com.jcloisterzone.figure.Meeple;
 import com.jcloisterzone.game.Game;
 import com.jcloisterzone.game.Snapshot;
 import com.jcloisterzone.game.capability.AbbeyCapability;
@@ -34,10 +28,6 @@ import com.jcloisterzone.game.capability.PlagueCapability;
 import com.jcloisterzone.game.capability.TowerCapability;
 import com.jcloisterzone.ui.Client;
 import com.jcloisterzone.ui.GameController;
-import com.jcloisterzone.ui.ImmutablePoint;
-import com.jcloisterzone.ui.animation.AnimationService;
-import com.jcloisterzone.ui.animation.FlierDiceRollAnimation;
-import com.jcloisterzone.ui.animation.ScoreAnimation;
 import com.jcloisterzone.ui.controls.ControlPanel;
 import com.jcloisterzone.ui.controls.chat.ChatPanel;
 import com.jcloisterzone.ui.grid.layer.AbbeyPlacementLayer;
@@ -69,7 +59,6 @@ public class MainPanel extends JPanel {
     private final GameView gameView;
     private final GameController gc;
     private final Game game;
-    private AnimationService animationService;
 
     private GridPanel gridPanel;
     private ControlPanel controlPanel;
@@ -86,9 +75,6 @@ public class MainPanel extends JPanel {
         this.chatPanel = chatPanel;
         gc.register(this);
 
-        animationService = new AnimationService();
-        animationService.start();
-
         setLayout(new BorderLayout());
     }
 
@@ -98,10 +84,6 @@ public class MainPanel extends JPanel {
 
     public ControlPanel getControlPanel() {
         return controlPanel;
-    }
-
-    public AnimationService getAnimationService() {
-        return animationService;
     }
 
     public void setShowFarmHints(boolean showFarmHints) {
@@ -169,8 +151,7 @@ public class MainPanel extends JPanel {
         //abstractare - zindex 100
         //tile placement 3
 
-        gridPanel.addLayer(new AnimationLayer(gridPanel, gc, animationService)); //zindex 800
-        animationService.setGridPanel(gridPanel);
+        gridPanel.addLayer(new AnimationLayer(gridPanel, gc)); //zindex 800
 
         placementHistoryLayer = new PlacementHistory(gridPanel, gc);
         gridPanel.addLayer(placementHistoryLayer, false);
@@ -211,42 +192,8 @@ public class MainPanel extends JPanel {
         gridPanel.clearActionDecorations();
     }
 
-
-    private Integer getScoreAnimationDuration() {
-        Integer duration = client.getConfig().getScore_display_duration();
-        return duration == null ? 10 : Math.max(duration, 1);
-    }
-
-    public void scored(Feature scoreable, Player player, String points, Class<? extends Meeple> meepleType, boolean finalScoring) {
-        Tile tile = scoreable.getTile();
-        Position pos = tile.getPosition();
-        ImmutablePoint offset = client.getResourceManager().getMeeplePlacement(tile, meepleType, scoreable.getLocation());
-        animationService.registerAnimation(new ScoreAnimation(
-            pos,
-            points,
-            offset,
-            player.getColors().getMeepleColor(),
-            finalScoring ? null : getScoreAnimationDuration()
-        ));
-    }
-
-    public void scored(Position pos, Player player, String points, boolean finalScoring) {
-        animationService.registerAnimation(new ScoreAnimation(
-            pos,
-            points,
-            new ImmutablePoint(50, 50),
-            player.getColors().getMeepleColor(),
-            finalScoring ? null : getScoreAnimationDuration()
-        ));
-    }
-
     @Subscribe
     public void scored(ScoreEvent ev) {
-        if (ev.getFeature() == null) {
-            scored(ev.getPosition(), ev.getTargetPlayer(), ev.getLabel(), ev.isFinal());
-        } else {
-            scored(ev.getFeature(), ev.getTargetPlayer(), ev.getLabel(), ev.getMeepleType(), ev.isFinal());
-        }
         repaint();
     }
 
@@ -261,11 +208,6 @@ public class MainPanel extends JPanel {
     @Subscribe
     public void onGoldChangeEvent(GoldChangeEvent ev) {
         gridPanel.repaint();
-    }
-
-    @Subscribe
-    public void flierRoll(FlierRollEvent ev) {
-        animationService.registerAnimation(new FlierDiceRollAnimation(ev.getPosition(), ev.getDistance()));
     }
 
     @Subscribe
